Move Batch to common
