+++ pyproject.toml
@@ -17,6 +17,7 @@ dependencies = [
     "jaxlib>=0.4.16",
     "flax>=0.8.3",
     "tensorboardX>=2.6.2.2",
+    "jaxtyping>=0.2.29",
 ]
 authors = [{ name = "Fang-Lin He", email = "fanglin.he.ms@gmail.com" }]
 classifiers = [
